requirements.txt
polars>=0.20.0
pyarrow>=14.0.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.31.0
python-dateutil>=2.8.0
numpy>=1.24.0
rich>=13.0.0
typer>=0.9.0
